Create README.md
At Fission Development, we're all about using apps and software to make the world a better place. We think creating tech shouldn't just be about making cool stuff; it should be about making things that help people, and everyone should be able to use them. We don't like apps that trick you into paying more and more money over time. Instead, we want to make apps and software that are free or that you just pay for once, without any sneaky extra costs.

Right now, we're still getting everything ready to go, but we're working hard. Our goal is to give everyone free or affordable options to replace the expensive apps and services out there today. We're excited about what we're building and can't wait to share it with you soon. We're on a mission to do good with technology, one step at a time.